Feat(util): handler now contains singing out and refreshing mechanisms
